requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.7
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
Django==4.2.7
executing==2.0.1
filelock==3.13.1
firebase-admin==6.2.0
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.45.0
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-speech==2.29.0
google-cloud-storage==2.13.0
google-cloud-texttospeech==2.13.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
graphviz==0.20.1
grpcio==1.59.2
grpcio-status==1.59.2
gTTS==2.5.3
h5py==3.10.0
httplib2==0.22.0
huggingface-hub==0.26.2
idna==3.4
imageai==3.0.3
instaloader==4.13.2
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.0
playsound==1.2.2
prompt-toolkit==3.0.43
proto-plus==1.22.3
protobuf==4.25.0
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydotplus==2.0.2
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
pypiwin32==223
pytesseract==0.3.10
python-dateutil==2.8.2
pyttsx3==2.98
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
six==1.16.0
sklearn-model==0.0.6
smmap==5.0.1
soupsieve==2.6
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tesseract==0.1.3
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tokenizers==0.20.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
transformers==4.46.1
twilio==9.0.4
typing_extensions==4.10.0
tzdata==2023.3
ultralytics==8.2.98
ultralytics-thop==2.0.6
uritemplate==4.1.1
urllib3==2.0.7
utils==1.0.2
virtualenv==20.26.5
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
